added README and requirements.txt, folder name renamed, removed unnecessary modules import, performed linting
