+++ Vaccine_Availability_Tracker/README.md
@@ -1,0 +1,12 @@
+## Vaccine Availability Tracker
+
+### In order to run this script
+1. [Download python IDLE](https://www.python.org/downloads/)
+2. Install requests package by running `pip install requests` in terminal.
+
+### How to use scripts
+Download the script and run "tracker_script.py".
+
+### Scripts description
+This script uses "Appointment Availability API" provided by API Setu.
+It will fetch response from API according to PIN and DATE, based on that it'll show available places.
