+++ Vaccine_Availability_Tracker/requirements.txt
@@ -1,0 +1,1 @@
+requests==2.23.0
